+++ packages/validation/src/index.ts
@@ -1,2 +1,3 @@
 export * from "./wedding";
 export * from "./vendor";
+export * from "./auth";